Rename 04_array/find_rules_hard.rb 04_array_hash/find_rules_hard.rb
