+++ src/04_array_hash/find_rules_hard.rb
